feat(系统管理): 角色优化
角色优化

+++ system/src/main/java/cn/cuiot/dmp/system/infrastructure/persistence/dao/RoleDao.java
@@ -191,7 +191,7 @@ public interface RoleDao {
      * @param roleId
      * @return
      */
-    List<AddMenuDto> getRoleMenu(Integer roleId);
+    List<AddMenuDto> getRoleMenu(Long roleId);
